WindTurbine Breaker added
Now aside of the relay activation, a break via chopper (doing DUTY -> 100%) will be applied.
This break will be removed after the voltage decreases below some arbitrary low voltage of 25V.
This break will not be activated again (even if conditions to do so are meet) until a 5 seconds timeout is reached.